+++ oldnyc/geocode/coders/gpt.py
@@ -6,7 +6,6 @@ import sys
 from oldnyc.geocode import grid
 from oldnyc.geocode.boroughs import guess_borough, point_to_borough
 from oldnyc.geocode.coders.coder_utils import get_lat_lng_from_geocode
-from oldnyc.geocode.coders.extended_grid import parse_street_ave
 from oldnyc.geocode.geocode_types import Coder, Locatable
 from oldnyc.geocode.geogpt.generate_batch import GptResponse
 from oldnyc.item import Item
@@ -80,7 +79,7 @@ class GptCoder(Coder):
         if boro != "Manhattan":
             return None
         try:
-            avenue, street = parse_street_ave(str1, str2)
+            avenue, street = grid.parse_street_ave(str1, str2)
             latlon = grid.code(avenue, street)
             if latlon:
                 self.n_grid += 1
